Suppression de commpte bancaire
FossilOrigin-Name: 3ea3b30a35a4c62f9636373cdec79b4b51fa5149

+++ include/class.compta_comptes_bancaires.php
@@ -53,10 +53,14 @@ class Garradin_Compta_Comptes_Bancaires extends Garradin_Compta_Comptes
 
     public function delete($id)
     {
-        $return = parent::delete($id);
-
         $db = Garradin_DB::getInstance();
+        if (!$db->simpleQuerySingle('SELECT 1 FROM compta_comptes_bancaires WHERE id = ?;', false, trim($id)))
+        {
+            throw new UserException('Ce compte n\'est pas un compte bancaire.');
+        }
+
         $db->simpleExec('DELETE FROM compta_comptes_bancaires WHERE id = ?;', trim($id));
+        $return = parent::delete($id);
 
         return $return;
     }
